requirements.txt
discord.py>=2.3,<3.0
python-dotenv>=1.0,<2.0
claude-agent-sdk>=0.1.0,<1.0
